refactor(functions): share one Stripe client and service-role gate
Three functions each constructed their own Stripe client with a
hand-written apiVersion string. They drifted: create/manage-subscription
moved to 2026-03-25.dahlia for the Payment Element migration while
stripe-webhook kept 2024-06-20, which governed nothing because that
function made no outbound calls. It will start mattering as soon as it
does, so collapse all three onto _shared/stripe.ts — one pinned version,
one factory, one explanation of why the SDK cast is needed.

The module also re-exports the SDK so it's imported in exactly one place,
and adds assertWebhookVersion(): the Dashboard endpoint decides what
version inbound payloads render at, nothing in this repo does, so a
mismatch between that and the code now shows up in the logs.

Also moves assertServiceRole out of cleanup-media into _shared, ahead of
purge-accounts needing the same gate. cleanup-media re-exports it so its
existing test suite keeps importing from the same place.
