+++ supabase/functions/create-subscription/index.ts
@@ -5,8 +5,8 @@
 // client_secret and calls `checkout.confirm()` — appearance stays ours,
 // Stripe owns creating the underlying Subscription + PaymentIntent.
 
-import Stripe from 'npm:stripe@20'
 import { createClient, type SupabaseClient } from '@supabase/supabase-js'
+import { makeStripe, type Stripe } from '../_shared/stripe.ts'
 
 const cors = {
   'Access-Control-Allow-Origin': '*',
@@ -124,12 +124,7 @@ export async function handler(req: Request, deps: Deps): Promise<Response> {
 }
 
 if (import.meta.main) {
-  const stripe = new Stripe(Deno.env.get('STRIPE_SECRET_KEY')!, {
-    // Preview version (see ui_mode note below) — not yet in this SDK's
-    // LatestApiVersion type.
-    apiVersion: '2026-03-25.dahlia' as Stripe.LatestApiVersion,
-    httpClient: Stripe.createFetchHttpClient()
-  })
+  const stripe = makeStripe()
 
   const admin = createClient(
     Deno.env.get('SUPABASE_URL')!,
